requirements.txt
kafka-python
colorama
requests
geoip2